Prion: Code review changes
