+++ ddp-workspace/projects/ddp-prion/src/app/components/redirect-join/redirect-join.component.ts
@@ -37,7 +37,7 @@ export class RedirectJoinComponent implements OnInit {
     }
   }
   private doAnalytics(): void {
-    this.analytics.emitCustomEvent('clickedCountMeIn', 'fromMainPage');
+    this.analytics.emitCustomEvent('clickedJoinUs', 'fromMainPage');
   }
 
 }
