+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/activityQuestion.component.ts
@@ -7,7 +7,6 @@ import { SubmissionManager } from '../../services/serviceAgents/submissionManage
 import { ConfigurationService } from '../../services/configuration.service';
 import { BehaviorSubject, combineLatest, Observable, Subject } from 'rxjs';
 import { delay, filter, map, shareReplay, startWith, takeUntil, tap } from 'rxjs/operators';
-import { BlockType } from '../../models/activity/blockType';
 
 @Component({
   selector: 'ddp-activity-question',
@@ -113,9 +112,6 @@ export class ActivityQuestionComponent implements OnInit, OnDestroy {
   }
 
   public getQuestionClass(block: ActivityQuestionBlock<any>): string {
-    if (block.blockType !== BlockType.Question) {
-      return '';
-    }
     return 'Question--' + block.questionType;
   }
