+++ ddp-workspace/projects/toolkit-prion/src/public-api.ts
@@ -5,7 +5,6 @@ export * from './lib/components/footer/prionFooter.component';
 export * from './lib/components/header/prionHeader.component';
 export * from './lib/components/activity-page/prionActivityPage.component';
 export * from './lib/components/workflow-start-activity/prionWorkflowStartActivity.component';
-export * from './lib/components/activity/prionActivity.component';
 export * from './lib/components/dashboard/prionDashboard.component';
 export * from './lib/components/error/prionError.component';
 export * from './lib/components/password/prionPassword.component';
