+++ ddp-workspace/projects/toolkit/src/lib/models/cookies/consentStatuses.ts
@@ -1,0 +1,5 @@
+export enum ConsentStatuses {
+  defaultAccept= 'default_accept',
+  defaultReject = 'default_reject',
+  managed = 'managed'
+}
