+++ ddp-workspace/projects/toolkit/src/lib/models/cookies/cookie.ts
@@ -1,0 +1,5 @@
+export interface Cookie {
+  name: string;
+  description: string;
+  duration: string;
+}
