+++ ddp-workspace/projects/toolkit/src/lib/models/cookies/cookies.ts
@@ -1,0 +1,10 @@
+import { CookiesTypes } from './cookiesType';
+import { Cookie } from './cookie';
+
+export interface Cookies {
+  data: Array<{
+    type: CookiesTypes;
+    actions: ['Accept', 'Reject'] | null;
+    list: Array<Cookie>;
+  }>;
+}
