+++ ddp-workspace/projects/toolkit/src/lib/models/cookies/cookiesConsentStatuses.ts
@@ -1,0 +1,1 @@
+export type CookiesConsentStatuses = 'default_accept' | 'default_reject' | 'managed';
