+++ ddp-workspace/projects/toolkit/src/lib/models/cookies/cookiesPreferences.ts
@@ -1,0 +1,7 @@
+import { CookiesConsentStatuses } from './cookiesConsentStatuses';
+
+export interface CookiesPreferences {
+  decision: boolean;
+  status: CookiesConsentStatuses;
+  cookies: {};
+}
