+++ ddp-workspace/projects/toolkit/src/lib/models/cookies/cookiesType.ts
@@ -1,0 +1,1 @@
+export type CookiesTypes = 'Functional' | 'Analytical';
